TST,DEP: Move deprecation tests that now test for exceptions.
Tests of deprecated indexing operations that now raise errors are moved
from test_deprecations.py to test_indexing.py.
