+++ numpy/core/tests/test_deprecations.py
@@ -115,213 +115,6 @@ class _DeprecationTestCase(object):
                         exceptions=tuple(), args=args, kwargs=kwargs)
 
 
-class TestFloatNonIntegerArgumentDeprecation(_DeprecationTestCase):
-    """
-    These test that ``DeprecationWarning`` is given when you try to use
-    non-integers as arguments to for indexing and slicing e.g. ``a[0.0:5]``
-    and ``a[0.5]``, or other functions like ``array.reshape(1., -1)``.
-
-    After deprecation, changes need to be done inside conversion_utils.c
-    in PyArray_PyIntAsIntp and possibly PyArray_IntpConverter.
-    In iterators.c the function slice_GetIndices could be removed in favor
-    of its python equivalent and in mapping.c the function _tuple_of_integers
-    can be simplified (if ``np.array([1]).__index__()`` is also deprecated).
-
-    As for the deprecation time-frame: via Ralf Gommers,
-
-    "Hard to put that as a version number, since we don't know if the
-    version after 1.8 will be 6 months or 2 years after. I'd say 2
-    years is reasonable."
-
-    I interpret this to mean 2 years after the 1.8 release. Possibly
-    giving a PendingDeprecationWarning before that (which is visible
-    by default)
-
-    """
-    message = "using a non-integer number instead of an integer " \
-              "will result in an error in the future"
-
-    def test_indexing(self):
-        a = np.array([[[5]]])
-
-        assert_raises(IndexError, lambda: a[0.0])
-        assert_raises(IndexError, lambda: a[0, 0.0])
-        assert_raises(IndexError, lambda: a[0.0, 0])
-        assert_raises(IndexError, lambda: a[0.0,:])
-        assert_raises(IndexError, lambda: a[:, 0.0])
-        assert_raises(IndexError, lambda: a[:, 0.0,:])
-        assert_raises(IndexError, lambda: a[0.0,:,:])
-        assert_raises(IndexError, lambda: a[0, 0, 0.0])
-        assert_raises(IndexError, lambda: a[0.0, 0, 0])
-        assert_raises(IndexError, lambda: a[0, 0.0, 0])
-        assert_raises(IndexError, lambda: a[-1.4])
-        assert_raises(IndexError, lambda: a[0, -1.4])
-        assert_raises(IndexError, lambda: a[-1.4, 0])
-        assert_raises(IndexError, lambda: a[-1.4,:])
-        assert_raises(IndexError, lambda: a[:, -1.4])
-        assert_raises(IndexError, lambda: a[:, -1.4,:])
-        assert_raises(IndexError, lambda: a[-1.4,:,:])
-        assert_raises(IndexError, lambda: a[0, 0, -1.4])
-        assert_raises(IndexError, lambda: a[-1.4, 0, 0])
-        assert_raises(IndexError, lambda: a[0, -1.4, 0])
-
-        # Test that the slice parameter deprecation warning doesn't mask
-        # the scalar index warning.
-        assert_raises(IndexError, lambda: a[0.0:, 0.0])
-        assert_raises(IndexError, lambda: a[0.0:, 0.0,:])
-
-    def test_valid_indexing(self):
-        a = np.array([[[5]]])
-        assert_not_deprecated = self.assert_not_deprecated
-
-        assert_not_deprecated(lambda: a[np.array([0])])
-        assert_not_deprecated(lambda: a[[0, 0]])
-        assert_not_deprecated(lambda: a[:, [0, 0]])
-        assert_not_deprecated(lambda: a[:, 0,:])
-        assert_not_deprecated(lambda: a[:,:,:])
-
-    def test_slicing(self):
-        a = np.array([[5]])
-
-        # start as float.
-        assert_raises(IndexError, lambda: a[0.0:])
-        assert_raises(IndexError, lambda: a[0:, 0.0:2])
-        assert_raises(IndexError, lambda: a[0.0::2, :0])
-        assert_raises(IndexError, lambda: a[0.0:1:2,:])
-        assert_raises(IndexError, lambda: a[:, 0.0:])
-        # stop as float.
-        assert_raises(IndexError, lambda: a[:0.0])
-        assert_raises(IndexError, lambda: a[:0, 1:2.0])
-        assert_raises(IndexError, lambda: a[:0.0:2, :0])
-        assert_raises(IndexError, lambda: a[:0.0,:])
-        assert_raises(IndexError, lambda: a[:, 0:4.0:2])
-        # step as float.
-        assert_raises(IndexError, lambda: a[::1.0])
-        assert_raises(IndexError, lambda: a[0:, :2:2.0])
-        assert_raises(IndexError, lambda: a[1::4.0, :0])
-        assert_raises(IndexError, lambda: a[::5.0,:])
-        assert_raises(IndexError, lambda: a[:, 0:4:2.0])
-        # mixed.
-        assert_raises(IndexError, lambda: a[1.0:2:2.0])
-        assert_raises(IndexError, lambda: a[1.0::2.0])
-        assert_raises(IndexError, lambda: a[0:, :2.0:2.0])
-        assert_raises(IndexError, lambda: a[1.0:1:4.0, :0])
-        assert_raises(IndexError, lambda: a[1.0:5.0:5.0,:])
-        assert_raises(IndexError, lambda: a[:, 0.4:4.0:2.0])
-        # should still get the DeprecationWarning if step = 0.
-        assert_raises(IndexError, lambda: a[::0.0])
-
-    def test_valid_slicing(self):
-        a = np.array([[[5]]])
-        assert_not_deprecated = self.assert_not_deprecated
-
-        assert_not_deprecated(lambda: a[::])
-        assert_not_deprecated(lambda: a[0:])
-        assert_not_deprecated(lambda: a[:2])
-        assert_not_deprecated(lambda: a[0:2])
-        assert_not_deprecated(lambda: a[::2])
-        assert_not_deprecated(lambda: a[1::2])
-        assert_not_deprecated(lambda: a[:2:2])
-        assert_not_deprecated(lambda: a[1:2:2])
-
-    def test_non_integer_argument_deprecations(self):
-        a = np.array([[5]])
-
-        assert_raises(TypeError, np.reshape, a, (1., 1., -1))
-        assert_raises(TypeError, np.reshape, a, (np.array(1.), -1))
-        assert_raises(TypeError, np.take, a, [0], 1.)
-        assert_raises(TypeError, np.take, a, [0], np.float64(1.))
-
-    def test_non_integer_sequence_multiplication(self):
-        # Numpy scalar sequence multiply should not work with non-integers
-        def mult(a, b):
-            return a * b
-
-        assert_raises(TypeError, mult, [1], np.float_(3))
-        self.assert_not_deprecated(mult, args=([1], np.int_(3)))
-
-    def test_reduce_axis_float_index(self):
-        d = np.zeros((3,3,3))
-        assert_raises(TypeError, np.min, d, 0.5)
-        assert_raises(TypeError, np.min, d, (0.5, 1))
-        assert_raises(TypeError, np.min, d, (1, 2.2))
-        assert_raises(TypeError, np.min, d, (.2, 1.2))
-
-
-class TestBooleanArgumentDeprecation(_DeprecationTestCase):
-    """This tests that using a boolean as integer argument/indexing is
-    deprecated.
-
-    This should be kept in sync with TestFloatNonIntegerArgumentDeprecation
-    and like it is handled in PyArray_PyIntAsIntp.
-    """
-    message = "using a boolean instead of an integer " \
-              "will result in an error in the future"
-
-    def test_bool_as_int_argument(self):
-        a = np.array([[[1]]])
-
-        assert_raises(TypeError, np.reshape, a, (True, -1))
-        assert_raises(TypeError, np.reshape, a, (np.bool_(True), -1))
-        # Note that operator.index(np.array(True)) does not work, a boolean
-        # array is thus also deprecated, but not with the same message:
-        assert_raises(TypeError, operator.index, np.array(True))
-        assert_raises(TypeError, np.take, args=(a, [0], False))
-        assert_raises(IndexError, lambda: a[False:True:True])
-        assert_raises(IndexError, lambda: a[False, 0])
-        assert_raises(IndexError, lambda: a[False, 0, 0])
-
-
-class TestArrayToIndexDeprecation(_DeprecationTestCase):
-    """This tests that creating an an index from an array is deprecated
-    if the array is not 0d.
-
-    This can probably be deprecated somewhat faster then the integer
-    deprecations. The deprecation period started with NumPy 1.8.
-    For deprecation this needs changing of array_index in number.c
-    """
-    message = "converting an array with ndim \> 0 to an index will result " \
-              "in an error in the future"
-
-    def test_array_to_index_deprecation(self):
-        # This drops into the non-integer deprecation, which is ignored here,
-        # so no exception is expected. The raising is effectively tested above.
-        a = np.array([[[1]]])
-
-        assert_raises(TypeError, operator.index, np.array([1]))
-        assert_raises(TypeError, np.reshape, a, (a, -1))
-        assert_raises(TypeError, np.take, a, [0], a)
-        
-        # Check slicing. Normal indexing checks arrays specifically.
-        assert_raises(IndexError, lambda: a[a:a:a])
-
-
-class TestNonIntegerArrayLike(_DeprecationTestCase):
-    """Tests that array likes, i.e. lists give a deprecation warning
-    when they cannot be safely cast to an integer.
-    """
-    message = "non integer \(and non boolean\) array-likes will not be " \
-              "accepted as indices in the future"
-
-    def test_basic(self):
-        a = np.arange(10)
-        assert_raises(IndexError, a.__getitem__, [0.5, 1.5])
-        assert_raises(IndexError, a.__getitem__, (['1', '2'],))
-
-        self.assert_not_deprecated(a.__getitem__, ([],))
-
-
-class TestMultipleEllipsisDeprecation(_DeprecationTestCase):
-    message = "an index can only have a single Ellipsis \(`...`\); replace " \
-              "all but one with slices \(`:`\)."
-
-    def test_basic(self):
-        a = np.arange(10)
-        assert_raises(IndexError, lambda : a[..., ...])
-        assert_raises(IndexError, a.__getitem__, ((Ellipsis,) * 2,))
-        assert_raises(IndexError, a.__getitem__, ((Ellipsis,) * 3,))
-
-
 class TestBooleanUnaryMinusDeprecation(_DeprecationTestCase):
     """Test deprecation of unary boolean `-`. While + and * are well
     defined, unary - is not and even a corrected form seems to have
